Fix all the compiler warnings on x64

+++ BIDebugEngine/BIDebugEngine/json.hpp
@@ -6678,7 +6678,7 @@ class basic_json
                                              thousands_sep);
                 std::fill(end, number_buffer.end(), '\0');
                 assert((end - number_buffer.begin()) <= len);
-                len = (end - number_buffer.begin());
+                len = static_cast<long>(end - number_buffer.begin());
             }
 
             // convert decimal point to '.'
